revamped the test layout
